+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
 ## [Unrealeased]
+### Fixed
+- Small bug in telling the user when checkM is not in system path
 
 ## [0.3.2]
 ### Fixed